+++ CHANGELOG.md
@@ -17,11 +17,19 @@ All user visible changes to this project will be documented in this file. This p
     - Media panel:
         - Redesigned participants modal. ([#127], [#122])
         - Dock buttons persistence. ([#159], [#137])
+        
+### Changed
+
+- UI:
+    - Chat page:
+        - Redesigned system messages. ([#161], [#158])
 
 [#122]: /../../issues/122
 [#127]: /../../pull/127
 [#137]: /../../issues/137
+[#158]: /../../issues/158
 [#159]: /../../pull/159
+[#161]: /../../pull/161